Assistant checkpoint: Update curriculum publishing workflow
Assistant generated file changes:
- templates/curriculum/edit.html: Remove publish and lock options from new curriculum form
- templates/curriculum/view.html: Add publish controls to curriculum view page
- routes.py: Add publish and lock routes

---

User prompt:

On curriculum/new there is a publish curriculum button and a lock curriculum button. I don't want those there. A new curriculum is private by default and should only be seen by the user who uploaded it. Once it is complete and has a curriculum details page, only from that page should the user be able to publish it for other users to use.